+++ src/main/java/com/alipay/global/api/response/ams/billing/AlipayPriceCreateResponse.java
@@ -58,7 +58,7 @@ public class AlipayPriceCreateResponse extends AlipayResponse {
   private Boolean active;
 
   /** The included quantity. */
-  private Integer includedQuantity;
+  private Long includedQuantity;
 
   /** The tiers mode. Maximum length: 16 characters. */
   private String tiersMode;
